+++ eveuniverse/models.py
@@ -1283,9 +1283,15 @@ class EveSolarSystem(EveUniverseEntityModel):
             destination: Other solar system to use in calculation
 
         Returns:
-            List of solar system objects incl. origin and destination or None if no route can be found (e.g. if one system is in WH space)
+            List of solar system objects incl. origin and destination
+            or None if no route can be found (e.g. if one system is in WH space)
         """
-        if self.is_w_space or destination.is_w_space:
+        if (
+            self.is_w_space
+            or destination.is_w_space
+            or self.is_trig_space
+            or destination.is_trig_space
+        ):
             return None
 
         path_ids = self._calc_route_esi(self.id, destination.id)
@@ -1304,7 +1310,8 @@ class EveSolarSystem(EveUniverseEntityModel):
             destination: Other solar system to use in calculation
 
         Returns:
-            Number of total jumps or None if no route can be found (e.g. if one system is in WH space)
+            Number of total jumps
+            or None if no route can be found (e.g. if one system is in WH space)
         """
         if (
             self.is_w_space