+++ eveuniverse/tests/test_models_3.py
@@ -979,7 +979,7 @@ class TestEveSolarSystemBulkWithSection(NoSocketsTestCase):
                     enabled_sections=EveSolarSystem.enabled_sections.planets
                 ).values_list("id", flat=True)
             ),
-            {30000142, 30001161, 30045339, 30045342, 31000005},
+            {30000142, 30001161, 30045339, 30045342, 31000005, 30000157},
         )
         self.assertEqual(
             set(EvePlanet.objects.values_list("id", flat=True)),
@@ -1004,7 +1004,7 @@ class TestEveSolarSystemBulkWithSection(NoSocketsTestCase):
                     enabled_sections=EveSolarSystem.enabled_sections.planets
                 ).values_list("id", flat=True)
             ),
-            {30000142, 30001161, 30045339, 30045342, 31000005},
+            {30000142, 30001161, 30045339, 30045342, 31000005, 30000157},
         )
         self.assertEqual(
             set(EvePlanet.objects.values_list("id", flat=True)),